# back-end/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Projeto Dispositivos com Categorias",
      "description": "API Dispositivos com Categorias swagger",
      "termsOfService": "http://localhost:3000/swagger-ui.html",
      "contact": {
        "email": "msbobsk8@gmail.com"
      },
      "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
...
    "schemes": [
      "http"
    ],
    "paths": {
...
      "/api/v1/device/create": {
        "post": {
          "tags": [
            "device"
          ],
          "summary": "Cadastro de dispositivos",
          "description": "Cadastra Dispositivos com de uma categoria.",
          "operationId": "Devices",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/device"
                },
                "example": {
                    "name": "Computador",
                    "color": "Preto",
                    "part_number": 1254,
                    "id_category": 1
                }
              }
            }
          },
          "responses": {}
        }
      },
...
      "/api/v1/category/create": {
        "post": {
          "tags": [
            "category"
          ],
          "summary": "Cadastro de categorias",
          "description": "Cadastra categorias.",
          "operationId": "Category",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/category"
                },
                "example": {
                    "name": "Ortopedia"
                }
              }
            }
          },
          "responses": {}
        }
      },
...
    },
    "components": {
      "schemas": {
        "device": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string"
            },
            "color": {
              "type": "string"
            },
            "part_number": {
              "type": "number"
            },
            "id_category": {
              "type": "number"
            }
          }
        },
        "category": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string"
              }
            }
          }
      }
    }
}